% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{HAVijSPLIT}
\alias{HAVijSPLIT}
\title{HAVijSPLIT}
\usage{
HAVijSPLIT(
  file,
  numeric_columns,
  independent_categorical_columns,
  dependent_categorical_columns,
  split_columns,
  path
)
}
\arguments{
\item{file}{A data frame containing the input data.}

\item{numeric_columns}{A numeric vector of column indices for numeric variables.}

\item{independent_categorical_columns}{A numeric vector of column indices for independent categorical variables.}

\item{dependent_categorical_columns}{A numeric vector of column indices for dependent categorical variables.}

\item{split_columns}{A numeric vector of column indices to split the data by.}

\item{path}{A character string of the destination path where the CSV files will be saved.}
}
\value{
For each split subset, a CSV file is saved with the test results.
}
\description{
This function processes the input data for each level of the specified split_columns.
For each split subset, it performs tests between numeric and categorical variables (based on normality),
correlations between numeric variables, and tests between categorical variables.
The results for each subset are saved as CSV files using a naming convention.
}
